[react-redux] RF: Separate reducers
