+++ src/redux/reducers/stats.js
@@ -1,0 +1,14 @@
+import {
+  STATS_CLEAR,
+  STATS_LOADED,
+} from '../types.js';
+
+export default function statsReducer(state = {}, action) {
+  switch (action.type) {
+    case STATS_LOADED:
+      return { stats: action.payload.issueCounts, isLoaded: true };
+    case STATS_CLEAR:
+      return { stats: [], isLoaded: false };
+    default: return state;
+  }
+}
